03 passport

+++ src/controllers/companiesController.ts
@@ -3,7 +3,7 @@ import { validationResult } from "express-validator";
 import dotenv from "dotenv";
 import slugify from "../util/slug";
 import Company from "../models/companies";
-
+import User from "../models/User";
 
 dotenv.config();
 
@@ -14,22 +14,25 @@ const addCompany = async (req: Request, res: Response) => {
             .status(422)
             .json({ message: "Validation error", errors: errors.array() });
     }
-   
+    const passportData = req.user as User;
+    const companyOwner = passportData.id;
+
     const name: string = req.body.name;
-    const logo: string = req.body.profilePhoto || "https://mom.rs" ;
-    const slug = slugify(name)
+    const logo: string =
+        req.body.profilePhoto ||
+        "https://mom.rs/wp-content/uploads/2016/10/test-logo.png";
+    const slug = slugify(name);
     try {
-        await Company.create({
+        const newCompany = await Company.create({
             name,
             logo,
-            slug
-        }).then((result) => {
-            console.log("Created new company", result);
-            res.status(201).json({"company": result});
-
+            slug,
+            companyOwner,
         });
+
+        res.status(201).json({ company: newCompany });
     } catch (err) {
-        console.log(err);
+        res.status(500).json({ message: "error creating Company" });
     }
 };
 
@@ -55,13 +58,18 @@ const getCompanybyId = async (req: Request, res: Response) => {
 };
 
 const deleteCompany = async (req: Request, res: Response) => {
-
     const id = Number(req.params.uid);
+    const passportData = req.user as User;
 
     try {
         const company = await Company.findOne({
             where: { id: id },
         });
+        if (company && passportData.id !== company.companyOwner) {
+            return res.status(403).json({
+                message: "Not authorized",
+            });
+        }
 
         if (company) {
             await Company.destroy({
@@ -75,8 +83,6 @@ const deleteCompany = async (req: Request, res: Response) => {
             res.status(400).json({ message: "Company doesn't exist" });
         }
     } catch (err) {
-        console.log(err);
-
         res.status(500).json({ message: "error deleting company" });
     }
 };
@@ -84,7 +90,6 @@ const deleteCompany = async (req: Request, res: Response) => {
 const updateCompany = async (req: Request, res: Response) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
-        console.log(errors);
         return res
             .status(422)
             .json({ message: "Validation error", errors: errors.array() });
@@ -95,6 +100,14 @@ const updateCompany = async (req: Request, res: Response) => {
         where: { id: id },
     });
 
+    const passportData = req.user as User;
+
+    if (company && passportData.id !== company.companyOwner) {
+        return res.status(403).json({
+            message: "Not authorized",
+        });
+    }
+
     const name = req.body.name || company?.name;
     const logo = req.body.logo || company?.logo;
     const slug = slugify(name) || company?.slug;
@@ -104,7 +117,7 @@ const updateCompany = async (req: Request, res: Response) => {
                 {
                     name,
                     logo,
-slug,
+                    slug,
                 },
                 {
                     where: {
@@ -115,7 +128,8 @@ slug,
             res.status(200).json({
                 message: `company ${id} updated`,
                 name,
-logo,                company,
+                logo,
+                company,
             });
         } else {
             res.status(404).json({ message: `Company not found` });
@@ -123,8 +137,6 @@ logo,                company,
     } catch (err) {
         res.status(500).json({ message: "Something went wrong" });
     }
-
-
 };
 
 export default {
